I was using objects of wrong types.
git-svn-id: https://svn.dev.java.net/svn/msv/trunk@1506 8cecbe3e-ef64-0410-a610-ece5afc49ac1

+++ xsdlib/src/com/sun/msv/datatype/xsd/ByteType.java
@@ -23,7 +23,7 @@ public class ByteType extends IntegerDerivedType {
 	public final static ByteType theInstance = new ByteType();
 	private ByteType() {
         super("byte",createRangeFacet(ShortType.theInstance,
-            new Short(Byte.MIN_VALUE), new Short(Byte.MAX_VALUE)));
+            new Byte(Byte.MIN_VALUE), new Byte(Byte.MAX_VALUE)));
     }
 	
 	final public XSDatatype getBaseType() {
